fixed the fix. THIS WILL NOT WORK
loading images, putting them on a canvas, and then trying to use
ctx.getImageData on it gives an error because of cross-site scripting.
these brushes will be on the same site later, so this won't be an error,
but it will not work while we're running on file:// for testing. we'll
have to set up local HTTP server for testing at that point. for now
reverting to the old method of brushes, with manually making each color